code: implement sortSmallFile function & pass func test
