+++ worker/src/test/scala/sorted_testFile3.txt
@@ -1,0 +1,10 @@
+04|China
+05|Brazil
+07|India
+08|Argentina
+16|Sudan
+20|Peru
+22|Niger
+23|Angola
+24|Mali
+28|Bolivia
