+++ worker/src/test/scala/sorted_testFile4.txt
@@ -1,0 +1,10 @@
+31|Tanzania
+32|Nigeria
+33|Venezuela
+34|Namibia
+35|Mozambique
+36|Pakistan
+37|Turkey
+38|Chile
+39|Zambia
+40|Myanmar
